Site picker is now called from fragment rather than activity
